backend/requirements.txt
django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.5.0
pillow==10.0.1
celery==5.3.4
redis==5.0.1
dj-database-url==2.1.0
python-decouple==3.8
django-storages==1.14.2
boto3==1.28.52
coverage==7.3.2
django-debug-toolbar==4.2.0
pywebpush==1.14.0
xlsxwriter==3.2.3
reportlab==4.2.0

# OCR
google-cloud-vision==3.4.3

# Monitoramento
sentry-sdk==1.40.5
django-prometheus==2.3.1

# WebSockets
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Assinatura Digital
requests==2.31.0
requests-toolbelt==1.0.0
python-dateutil==2.8.2 